make max age seconds configurable
